+++ crates/store/src/db/mod.rs
@@ -7,7 +7,7 @@ use anyhow::Context;
 use diesel::{Connection, RunQueryDsl, SqliteConnection};
 use miden_lib::utils::Serializable;
 use miden_node_proto::{
-    domain::account::{AccountInfo, AccountSummary},
+    domain::account::{AccountInfo, AccountSummary, NetworkAccountPrefix},
     generated as proto,
 };
 use miden_objects::{
@@ -484,6 +484,28 @@ impl Db {
     ) -> Result<(Vec<NoteRecord>, Page)> {
         self.transact("unconsumed network notes", move |conn| {
             models::queries::unconsumed_network_notes(conn, page)
+        })
+        .await
+    }
+
+    /// Loads the network notes for an account that are unconsumed by a specified block number.
+    /// Pagination is used to limit the number of notes returned.
+    pub(crate) async fn select_unconsumed_network_notes_for_account(
+        &self,
+        network_account_id_prefix: NetworkAccountPrefix,
+        block_num: BlockNumber,
+        page: Page,
+    ) -> Result<(Vec<NoteRecord>, Page)> {
+        // Network notes sent to a specific account have their tags set to the prefix of the target
+        // account ID. So we can convert the ID prefix into a note tag to query the notes for a
+        // given account.
+        self.transact("unconsumed network notes for account", move |conn| {
+            models::queries::select_unconsumed_network_notes_by_tag(
+                conn,
+                network_account_id_prefix.into(),
+                block_num,
+                page,
+            )
         })
         .await
     }